feat: update TestimonialCarousel component with new layout and styles; add gray_20 token
